Updated package.json and gitignore files, changed project name, description, type was set to module, author name was properly set, license changed, organized files into js, css and assets/images folders and made sure everything was working.
